+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/guessing_card/card_association_result/CardAssociationResultErrorActivity.java
@@ -10,7 +10,7 @@ import com.mercadolibre.android.ui.widgets.MeliButton;
 import com.mercadopago.android.px.R;
 import com.mercadopago.android.px.core.internal.MercadoPagoCardStorage;
 import com.mercadopago.android.px.internal.features.guessing_card.GuessingCardActivity;
-import com.mercadopago.android.px.internal.util.StatusBarDecorator;
+import com.mercadopago.android.px.internal.viewmodel.StatusBarDecorator;
 import com.mercadopago.android.px.tracking.internal.views.CardAssociationResultViewTrack;
 
 public class CardAssociationResultErrorActivity extends AppCompatActivity {
